[PATCH 3/7] mkimage: account for space for trampolines earlier
In particular we will want to do this before assigning space for .bss,
otherwise the trampoline will potentially overlap the start of .bss and might
be cleared before it is used.

I'm in two minds about whether this is fixing and actual (perhaps latent) bug
or not. AFAICT none of the existing targets which need trampolines require a
.bss assignment here (e.g. they generate PE images which means the .bss gets
taken care of at load time).

Apart from duplicating the surrounding if statement this is pure code motion.

Signed-off-by: Ian Campbell <address@hidden>