[PATCH 6/7] mkimage: support images which require full relocation at	mkimage time.
To do so support the specification of a "target_address" (which must be 0 for
image types which do not have this flag) and a new image flag to signal the
need for absolute relocation. If this flag is present then image->link_addr is
the default target.

Account for the target address when relocating.

Fabricate correct addresses for absolute symbols relating to the bss
("__bss_start" and "_end").

This functionality is not yet used by any image type and is connected only to
grub-mkimage ("-T option) and not yet to grub-install.

Signed-off-by: Ian Campbell <address@hidden>